cotxox/src/main/java/edu/craptocraft/cotxox/carrera/Carrera.java
package edu.craptocraft.cotxox.carrera;

public class Carrera {

    private String tarjetaCredito;
    private String origen;
    private String destino;
    private double distancia = 0d;
    private int tiempoEsperado;

    public Carrera(String tarjetaCredito) {
        this.tarjetaCredito = tarjetaCredito;
    }

    public String getTarjetaCredito() {
        return this.tarjetaCredito;
    }

    public void setOrigen(String origen) {
        this.origen = origen;
    }

    public String getOrigen() {
        return this.origen;
    }

    public void setDestino(String destino) {
        this.destino = destino;
    }

    public String getDestino() {
        return this.destino;
    }

    public void setDistancia(double distancia) {
        this.distancia = distancia;
    }

    public double getDistancia() {
        return this.distancia;
    }
}
